requirements.txt
luigi
pandas
numpy
SQLAlchemy
psycopg2-binary
loguru
python-dotenv
playwright
beautifulsoup4
tqdm